requirements.txt
aiofiles==0.8.0
aiohttp==3.8.1
aiohttp-swagger==1.0.16
aiosignal==1.2.0
asgiref==3.5.0
async-lru==1.0.2
async-timeout==4.0.2
attrs==21.4.0
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.12
cisco-gnmi==1.0.15
colorama==0.4.4
cryptography==36.0.1
dill==0.3.4
distro==1.7.0
Django==4.0.3
django-crispy-forms==1.14.0
frozenlist==1.3.0
future==0.18.2
genie==22.2
genie.libs.clean==22.2
genie.libs.conf==22.2
genie.libs.filetransferutils==22.2
genie.libs.health==22.2
genie.libs.ops==22.2
genie.libs.parser==22.2
genie.libs.sdk==22.2
gitdb==4.0.9
GitPython==3.1.18
grpcio==1.44.0
idna==3.3
importlib-metadata==4.11.2
Jinja2==3.0.3
jsonpickle==2.1.0
junit-xml==1.9
lxml==4.8.0
MarkupSafe==2.1.0
multidict==6.0.2
mypy-extensions==0.4.3
netaddr==0.8.0
nornir==3.2.0
nornir-scrapli==2022.1.30
nornir-utils==0.1.2
ntc-templates==3.0.0
packaging==21.3
pathspec==0.9.0
Pillow==9.0.1
prettytable==3.2.0
protobuf==3.19.4
psutil==5.9.0
pyats==22.2
pyats.aereport==22.2
pyats.aetest==22.2
pyats.async==22.2
pyats.connections==22.2
pyats.datastructures==22.2
pyats.easypy==22.2
pyats.kleenex==22.2
pyats.log==22.2
pyats.reporter==22.2
pyats.results==22.2
pyats.tcl==22.2
pyats.topology==22.2
pyats.utils==22.2
pycparser==2.21
pyftpdlib==1.5.6
pyparsing==3.0.7
python-engineio==3.14.2
python-socketio==4.6.1
PyYAML==6.0
requests==2.27.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
scrapli==2022.1.30.post1
scrapli-cfg==2022.1.30
scrapli-community==2022.1.30
scrapli-netconf==2022.1.30
six==1.16.0
smmap==5.0.0
sqlparse==0.4.2
textfsm==1.1.2
tftpy==0.8.0
tqdm==4.63.0
typing-extensions==3.10.0.2
unicon==22.2
unicon.plugins==22.2
urllib3==1.26.8
wcwidth==0.2.5
xmltodict==0.12.0
yamllint==1.26.3
yarl==1.7.2
zipp==3.7.0
